Rename channel classes for readability
