+++ framework/src/controller/channels/base_channel.js
